[COM-2163] Add warning that CASOpcDaWrapper.exe is opening CASOpcDaWrapper.config, not CASOpcDaWrapper.exe.config the following message is added:
---------------------------------------------
WARNING: The configuration (.config) file is missing. Usually CommServer OPC server executable (CASOpcDaWrapper.exe) is trying to open: 'C:\Program Files\CAS\CommServer\CASOpcDaWrapper.config' but this file does not exist. The file that exists in the directory is: 'C:\Program Files\CAS\CommServer\CASOpcDaWrapper.exe.config'.

EXPLANATION:This situation appears someties in Windows Server operating system, somehow the operating system is trying to find .config file (not .exe.config).

SOLUTION: The solution is to create of a copy of the 'CASOpcDaWrapper.exe.config' file and rename the copy 'CASOpcDaWrapper.config'.

Note: Before reconnection to OPC server make sure that CASOpcDaWrapper.exe is not appear on the task list (in Windows Task Manager). If it is exist please kill the process before reconnection.
--------------------------------------------
Issue #: 2163